day 08: use manual implementation of CRT
This actually passes all the tests, although it explodes with an out of
bounds multiply in debug mode.

While the "sieving" method is not asymptotically fast, it's actually
conceptually quite simple, so I went with that one.  It'll also let me
fairly easily handle not-coprime moduli, since the logic sounds like
it'll still work iff there is a solution.