day 17: only track last movement direction
This lets us avoid keeping so many different ways-to-get-to-a-coordinate
in our HashMap of distances; the only thing that affects the choices is
how many spaces we've moved in the last direction.  It doesn't matter
what direction we were going when we turned.

This brings the personal-input test runtime down from 996 seconds to 58
seconds, for a 17x speedup.  And from ~2GB of memory to just under 300MB
of memory.